refactor: quick dist speed
